requirements.txt
beautifulsoup4==4.5.3
blinker==1.4
bs4==0.0.1
click==6.7
factory-boy==2.8.1
Faker==0.7.7
Flask==0.12
Flask-Login==0.4.0
Flask-OAuthlib==0.9.3
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
Flask-WTF==0.13.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
oauthlib==2.0.1
psycopg2==2.6.2
python-dateutil==2.6.0
python-dotenv==0.6.2
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.15
WTForms==2.1
